rePercussion/requirements.txt
Flask
spotipy
python-dotenv
requests
gunicorn


